+++ test/dicer-malformed-header.test.js
@@ -1,14 +1,19 @@
-const Dicer = require('../deps/dicer/lib/Dicer')
-const { expect } = require('chai')
+'use strict'
 
-describe('dicer-malformed-header', () => {
-  it('should gracefully handle headers with leading whitespace', done => {
+const { test } = require('tap')
+const Dicer = require('../deps/dicer/lib/Dicer')
+
+test('dicer-malformed-header', t => {
+  t.plan(1)
+
+  t.test('should gracefully handle headers with leading whitespace', t => {
+    t.plan(3)
     const d = new Dicer({ boundary: '----WebKitFormBoundaryoo6vortfDzBsDiro' })
 
     d.on('part', function (p) {
       p.on('header', function (header) {
-        expect(header).has.property(' content-disposition')
-        expect(header[' content-disposition']).to.be.eql(['form-data; name="bildbeschreibung"'])
+        t.hasProp(header, ' content-disposition')
+        t.strictSame(header[' content-disposition'], ['form-data; name="bildbeschreibung"'])
       })
       p.on('data', function (data) {
       })
@@ -16,7 +21,7 @@ describe('dicer-malformed-header', () => {
       })
     })
     d.on('finish', function () {
-      done()
+      t.pass()
     })
 
     d.write(Buffer.from('------WebKitFormBoundaryoo6vortfDzBsDiro\r\n Content-Disposition: form-data; name="bildbeschreibung"\r\n\r\n\r\n------WebKitFormBoundaryoo6vortfDzBsDiro--'))
